move GitignoreStack to its own file
